feat: separate client config
