+++ scripts/start.ts
@@ -15,8 +15,8 @@ import webpack from "webpack";
 import WebpackDevServer from "webpack-dev-server";
 
 import createDevServerConfig from "../config/webpackDevServer.config";
+import config from "../src/core/build/config";
 import createWebpackConfig from "../src/core/build/createWebpackConfig";
-import config, { createClientEnv } from "../src/core/common/config";
 
 // tslint:disable: no-console
 
@@ -63,9 +63,7 @@ choosePort(HOST, PORT)
     const protocol = "http";
     const appName = "Talk";
     const urls = prepareUrls(protocol, HOST, port);
-    const webpackConfig = createWebpackConfig({
-      env: createClientEnv(config),
-    });
+    const webpackConfig = createWebpackConfig(config);
     // Create a webpack compiler that is configured with custom messages.
     const compiler = createCompiler(webpack, webpackConfig, appName, urls);
     // Serve webpack assets generated by the compiler over a web sever.
