+++ src/core/server/graph/common/subscriptions/pubsub.ts
@@ -1,5 +1,5 @@
 import { RedisPubSub } from "graphql-redis-subscriptions";
-import { Config } from "talk-common/config";
+import { Config } from "talk-server/config";
 import { createRedisClient } from "talk-server/services/redis";
 
 export function createPubSub(config: Config): RedisPubSub {
